+++ poptorch/include/poptorch/PoplarExecutable.hpp
@@ -60,6 +60,7 @@ public:
   bool isAttachedToDevice() const;
 
   const Compiler &getCompiler() const { return _compiler; }
+  Compiler &getCompiler() { return _compiler; }
 
 private:
   poptorch::Compiler _compiler;