+++ scripts/PopParse.py
@@ -37,7 +37,6 @@ logging.basicConfig(level=logging_level)
 # List of SessionOptions attributes PopTorch decided to not support
 options_not_handled = [
     "developerSettings",
-    "enableLoadAndOffloadRNGState",
     "prefetchBufferingDepthMap",
     "matmulOptions",
     "tensorLocationSettingsOverride",